Change link to mailing lists in README.md (#7)
The current link to "mailing lists" takes you to the top level Groups.IO list. It is not obvious from here there are any other mailing lists for OpenSSF. It took me an embarrassingly long time to figure out where the actual list of mailing lists existed. I believe this is a better, more intuitive link.

Co-authored-by: John Mertic <jmertic@linuxfoundation.org>