Create OpenSSF Policies and Procedures.md
Add the OpenSSF Policies and Procedures document to the foundation GH repo for a more accessbile and public location. This document was approved by the governing board in December 2023. Please note that I have left off the Background section of the document - if needed, I can add a blurb at the bottom and link to the full document. 

Signed-off-by: Naomi Washington <70667788+Naomi-Wash@users.noreply.github.com>